Create README.md
The Library management system (LMS) is a robust desktop application built using java swing. It is designed to manage the various operations 
of a library efficiency providing a comprehensive solution for libraries and users. this application allows for the streamlined management
of books, users. this registration and transaction handling makin it an essential tool for modern libraries